// frontend-new/models/socio-cultural/socio.ts
export interface SocioCulturalStats {
    winningParty: string;
    wahlkreisName: string;
    type: string;
    svbInsgesamt: number;
    svbLandwFischerei: number;
    svbProduzGewerbe: number;
    svbHandelGastVerkehr: number;
    svbDienstleister: number;
    alterUnter18: number;
    alter1824: number;
    alter2534: number;
    alter3559: number;
    alter6074: number;
    alter75Plus: number;
    alqFrauen: number;
    alq1524: number;
    alq5564: number;
    alqInsgesamt: number;
    alqMaenner: number;
}